--- lexer/base/base.c
@@ -1,8 +1,0 @@
-#include "base.h"
-
-void skip_spaces(char *line) {
-    int i = 0;
-    while (line[i] == ' ' || line[i] == '\t') {
-        i++;
-    }
-}
